+++ src/apps/mesoscale-explorer/ui/states.tsx
@@ -84,6 +84,8 @@ function adjustPluginProps(ctx: PluginContext) {
                     blurDepthBias: 0.5,
                     resolutionScale: 1,
                     color: Color(0x000000),
+                    includeTransparent: true,
+                    transparentThreshold: 0.6
                 }
             },
             shadow: {