+++ src/mol-gl/shader/outlines.frag.ts
@@ -40,7 +40,7 @@ float getDepthOpaque(const in vec2 coords) {
 
 float getDepthTransparent(const in vec2 coords) {
     #ifdef dTransparentOutline
-        return unpackRGBAToDepth(texture2D(tDepthTransparent, coords));
+        return unpackRGBAToDepthWithAlpha(texture2D(tDepthTransparent, coords)).x;
     #else
         return 1.0;
     #endif